+++ ndk-build/src/ndk.rs
@@ -4,6 +4,10 @@ use std::collections::HashMap;
 use std::path::{Path, PathBuf};
 use std::process::Command;
 
+/// The default password used when creating the default `debug.keystore` via
+/// [`Ndk::debug_key`]
+pub const DEFAULT_DEV_KEYSTORE_PASSWORD: &str = "android";
+
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub struct Ndk {
     sdk_path: PathBuf,
@@ -384,7 +388,8 @@ impl Ndk {
 
     pub fn debug_key(&self) -> Result<Key, NdkError> {
         let path = self.android_user_home()?.join("debug.keystore");
-        let password = "android".to_string();
+        let password = DEFAULT_DEV_KEYSTORE_PASSWORD.to_owned();
+
         if !path.exists() {
             let mut keytool = self.keytool()?;
             keytool